refactoring: moved altitude and identity decoding to dedicated class
