+++ src/main/java/de/serosystems/lib1090/msgs/adsb/AirbornePositionV0Msg.java
@@ -2,6 +2,7 @@ package de.serosystems.lib1090.msgs.adsb;
 
 import de.serosystems.lib1090.Position;
 import de.serosystems.lib1090.cpr.CPREncodedPosition;
+import de.serosystems.lib1090.decoding.Altitude;
 import de.serosystems.lib1090.exceptions.BadFormatException;
 import de.serosystems.lib1090.exceptions.UnspecifiedFormatError;
 import de.serosystems.lib1090.msgs.PositionMsg;
@@ -336,57 +337,10 @@ public class AirbornePositionV0Msg extends ExtendedSquitter implements Serializa
 		return altitude_available;
 	}
 
-	/**
-	 * This method converts a gray code encoded int to a standard decimal int
-	 * @param gray gray code encoded int of length bitlength
-	 *        bitlength bitlength of gray code
-	 * @return radix 2 encoded integer
-	 */
-	private static int grayToBin(int gray, int bitlength) {
-		int result = 0;
-		for (int i = bitlength-1; i >= 0; --i)
-			result = result|((((0x1<<(i+1))&result)>>>1)^((1<<i)&gray));
-		return result;
-	}
-
-	public static Integer decodeAltitude(short altitude_encoded) {
-		boolean Qbit = (altitude_encoded&0x10)!=0;
-		int N;
-		if (Qbit) { // altitude reported in 25ft increments
-			N = (altitude_encoded&0xF) | ((altitude_encoded&0xFE0)>>>1);
-			return 25*N-1000;
-		}
-		else { // altitude is above 50175ft, so we use 100ft increments
-
-			// it's decoded using the Gillham code
-			int C1 = (0x800&altitude_encoded)>>>11;
-			int A1 = (0x400&altitude_encoded)>>>10;
-			int C2 = (0x200&altitude_encoded)>>>9;
-			int A2 = (0x100&altitude_encoded)>>>8;
-			int C4 = (0x080&altitude_encoded)>>>7;
-			int A4 = (0x040&altitude_encoded)>>>6;
-			int B1 = (0x020&altitude_encoded)>>>5;
-			int B2 = (0x008&altitude_encoded)>>>3;
-			int D2 = (0x004&altitude_encoded)>>>2;
-			int B4 = (0x002&altitude_encoded)>>>1;
-			int D4 = (0x001&altitude_encoded);
-
-			// this is standard gray code
-			int N500 = grayToBin(D2<<7|D4<<6|A1<<5|A2<<4|A4<<3|B1<<2|B2<<1|B4, 8);
-
-			// 100-ft steps must be converted
-			int N100 = grayToBin(C1<<2|C2<<1|C4, 3)-1;
-			if (N100 == 6) N100=4;
-			if (N500%2 != 0) N100=4-N100; // invert it
-
-			return -1200+N500*500+N100*100;
-		}
-	}
-
 	@Override
 	public Integer getAltitude() {
 		if (!altitude_available) return null;
-		return decodeAltitude(altitude_encoded);
+		return Altitude.decode12BitAltitude(altitude_encoded);
 	}
 
 	@Override
