+++ src/main/java/de/serosystems/lib1090/msgs/adsb/IdentificationMsg.java
@@ -1,5 +1,6 @@
 package de.serosystems.lib1090.msgs.adsb;
 
+import de.serosystems.lib1090.decoding.Identification;
 import de.serosystems.lib1090.exceptions.BadFormatException;
 import de.serosystems.lib1090.exceptions.UnspecifiedFormatError;
 import de.serosystems.lib1090.msgs.modes.ExtendedSquitter;
@@ -34,31 +35,6 @@ public class IdentificationMsg extends ExtendedSquitter implements Serializable 
 
 	private byte emitter_category;
 	private byte[] identity;
-
-	/**
-	 * Maps ADS-B encoded to readable characters
-	 * @param digit encoded digit
-	 * @return readable character
-	 */
-	private static char mapChar (byte digit) {
-		if (digit>0 && digit<27) return (char) ('A'+digit-1);
-		else if (digit>47 && digit<58) return (char) ('0'+digit-48);
-		else return ' ';
-	}
-
-	/**
-	 * Maps ADS-B encoded to readable characters
-	 * @param digits array of encoded digits
-	 * @return array of decoded characters
-	 */
-	public static char[] mapChar (byte[] digits) {
-		char[] result = new char[digits.length];
-
-		for (int i=0; i<digits.length; i++)
-			result[i] = mapChar(digits[i]);
-
-		return result;
-	}
 
 	/** protected no-arg constructor e.g. for serialization with Kryo **/
 	protected IdentificationMsg() { }
@@ -97,28 +73,7 @@ public class IdentificationMsg extends ExtendedSquitter implements Serializable 
 		emitter_category = (byte) (msg[0] & 0x7);
 
 		// extract identity
-		identity = decodeAircraftIdentification(msg);
-	}
-
-	public static byte[] decodeAircraftIdentification(byte[] msg) {
-		byte[] identity = new byte[8];
-
-		int byte_off, bit_off;
-		for (int i=8; i>=1; i--) {
-			// calculate offsets
-			byte_off = (i*6)/8; bit_off = (i*6)%8;
-
-			// char aligned with byte?
-			if (bit_off == 0) identity[i-1] = (byte) (msg[byte_off]&0x3F);
-			else {
-				++byte_off;
-				identity[i-1] = (byte) (msg[byte_off]>>>(8-bit_off)&(0x3F>>>(6-bit_off)));
-				// should we add bits from the next byte?
-				if (bit_off < 6) identity[i-1] |= msg[byte_off-1]<<bit_off&0x3F;
-			}
-		}
-
-		return identity;
+		identity = Identification.decodeAircraftIdentification(msg);
 	}
 
 	/**
@@ -132,56 +87,7 @@ public class IdentificationMsg extends ExtendedSquitter implements Serializable 
 	 * @return the call sign as 8 characters array
 	 */
 	public char[] getIdentity() {
-		return mapChar(identity);
-	}
-
-	/**
-	 * @param type_code format type code of identity message
-	 * @param emitter_category reported emitter category
-	 * @return a textual description of the emitter's category according to DO-260B
-	 */
-	public static String categoryDescription(byte type_code, byte emitter_category) {
-		// category descriptions according
-		// to the ADS-B specification
-		String[][] categories = {{
-				"No ADS-B Emitter Category Information",
-				"Light (< 15500 lbs)",
-				"Small (15500 to 75000 lbs)",
-				"Large (75000 to 300000 lbs)",
-				"High Vortex Large (aircraft such as B-757)",
-				"Heavy (> 300000 lbs)",
-				"High Performance (> 5g acceleration and 400 kts)",
-				"Rotorcraft"
-		},{
-				"No ADS-B Emitter Category Information",
-				"Glider / sailplane",
-				"Lighter-than-air",
-				"Parachutist / Skydiver",
-				"Ultralight / hang-glider / paraglider",
-				"Reserved",
-				"Unmanned Aerial Vehicle",
-				"Space / Trans-atmospheric vehicle",
-		},{
-				"No ADS-B Emitter Category Information",
-				"Surface Vehicle – Emergency Vehicle",
-				"Surface Vehicle – Service Vehicle",
-				"Point Obstacle (includes tethered balloons)",
-				"Cluster Obstacle",
-				"Line Obstacle",
-				"Reserved",
-				"Reserved"
-		},{
-				"Reserved",
-				"Reserved",
-				"Reserved",
-				"Reserved",
-				"Reserved",
-				"Reserved",
-				"Reserved",
-				"Reserved"
-		}};
-
-		return categories[4-type_code][emitter_category];
+		return Identification.mapChar(identity);
 	}
 
 	/**
@@ -189,7 +95,7 @@ public class IdentificationMsg extends ExtendedSquitter implements Serializable 
 	 *         the ADS-B message format specification
 	 */
 	public String getCategoryDescription() {
-		return categoryDescription(getFormatTypeCode(), emitter_category);
+		return Identification.categoryDescription(getFormatTypeCode(), emitter_category);
 	}
 
 	@Override
