+++ src/main/java/de/serosystems/lib1090/msgs/tisb/IdentificationMsg.java
@@ -7,8 +7,8 @@ import de.serosystems.lib1090.msgs.modes.ExtendedSquitter;
 import java.io.Serializable;
 import java.util.Arrays;
 
-import static de.serosystems.lib1090.msgs.adsb.IdentificationMsg.categoryDescription;
-import static de.serosystems.lib1090.msgs.adsb.IdentificationMsg.mapChar;
+import static de.serosystems.lib1090.decoding.Identification.categoryDescription;
+import static de.serosystems.lib1090.decoding.Identification.mapChar;
 
 /*
  *  This file is part of de.serosystems.lib1090.
